Merged all migrations

+++ database/migrations/2024_02_23_163242_create_conditions_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('conditions', function (Blueprint $table) {
             $table->id();
-            $table->integer('criteriaId');
+            $table->unsignedBigInteger('criteriaId');
             $table->integer('min');
             $table->integer('max');
             $table->timestamps();
